Extract settings list code to 'SettingsListFragment'
